+++ src/server/oph_io_server_query_blocks.c
@@ -1378,7 +1378,7 @@ int _oph_io_server_query_load_from_file(oph_metadb_db_row ** meta_db, oph_iostor
 		return OPH_IO_SERVER_MEMORY_ERROR;
 	}
 
-	int dim_unlim = 0;
+	int dim_unlim = -1;
 	char *dim_unlimited = hashtbl_get(query_args, OPH_QUERY_ENGINE_LANG_ARG_DIM_UNLIM);
 	if (dim_unlimited)
 		dim_unlim = (int) strtol(dim_unlimited, NULL, 10);